fix: Remove unused alter table statements
